Rename asset viewer folder
