+++ web/src/lib/components/asset-viewer/photo-viewer.svelte
